* commit '3cad2709877f48d6f797d20a38023402052f4315':
  Dto Changes due to Db change
  Dto update due to Edmx change
  Fixed enums.

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Contact/CountryDto.cs
@@ -11,7 +11,7 @@ namespace PS.Mothership.Core.Common.Dto.Contact
     public class CountryDto
     {
         [DataMember]
-        public long CountryKey { get; set; }
+        public int CountryKey { get; set; }
 
         [DataMember]
         public string CountryName { get; set; }
